+++ src/main/java/kr/hhplus/be/server/domain/payment/dto/PaymentResult.java
@@ -3,10 +3,12 @@ package kr.hhplus.be.server.domain.payment.dto;
 import kr.hhplus.be.server.domain.payment.Payment;
 import kr.hhplus.be.server.domain.payment.PaymentStatus;
 
+import java.math.BigDecimal;
+
 public record PaymentResult(
         Long id,
         Long orderId,
-        int paymentPrice,
+        BigDecimal paymentPrice,
         PaymentStatus paymentStatus
 ) {
