+++ src/main/java/kr/hhplus/be/server/interfaces/api/point/request/PointChargeRequest.java
@@ -2,9 +2,11 @@ package kr.hhplus.be.server.interfaces.api.point.request;
 
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record PointChargeRequest(
         @NotNull
         Long userId,
-        int amount
+        BigDecimal amount
 ) {
 }
